Move scancode.utils to commoncode.cliutils #2233
These utilities are used by extractcode and scancode CLI.
